+++ IdentityCommand.SIA/Public/Get-SIAMFAKey.ps1
@@ -1,8 +1,13 @@
 # .ExternalHelp IdentityCommand.SIA-help.xml
-function Get-SIAMFACachingKey {
+function Get-SIAMFAKey {
     [CmdletBinding()]
     param(
-
+        [parameter(
+            Mandatory = $false,
+            ValueFromPipelinebyPropertyName = $true
+        )]
+        [ValidateSet('openssh', 'ppk')]
+        [String]$format
     )
 
     BEGIN { }#begin
@@ -10,6 +15,10 @@ function Get-SIAMFACachingKey {
     PROCESS {
 
         $URI = "$($ISPSSSession.tenant_url)/api/ssh/sso/key"
+
+        if ($PSBoundParameters.ContainsKey('format')) {
+            $URI = "$URI`?format=$format"
+        }
 
         #Send Request
         $result = Invoke-IDRestMethod -Uri $URI -Method GET